server/socialmedia/requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
Pillow
